requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
langdetect
pypdf
numpy
tenacity>=8.0.0